integreted light_sensor_reader in InGameFragment
