+++ app/src/main/java/com/uni/gruppenphaseandroid/InGameFragment.java
@@ -1,6 +1,12 @@
 package com.uni.gruppenphaseandroid;
 
+import android.content.Context;
+import android.hardware.Sensor;
+import android.hardware.SensorEvent;
+import android.hardware.SensorEventListener;
+import android.hardware.SensorManager;
 import android.os.Bundle;
+import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -19,10 +25,12 @@ import com.uni.gruppenphaseandroid.manager.GameManager;
 import com.uni.gruppenphaseandroid.playingfield.FigureManager;
 import com.uni.gruppenphaseandroid.playingfield.PlayingField;
 
-public class InGameFragment extends Fragment {
+public class InGameFragment extends Fragment implements SensorEventListener {
     FigureManager figureManager;
     private Client websocketClient;
     private final Gson gson = new Gson();
+    private SensorManager sensorManager;
+    private Sensor sensor;
 
     @Override
     public View onCreateView(
@@ -42,6 +50,7 @@ public class InGameFragment extends Fragment {
         GameManager.getInstance().setWebSocketClient(((MainActivity) getContext()).getWebsocketClient());
 
 
+
         view.findViewById(R.id.bttn_leave_game).setOnClickListener(view1 -> {
             websocketClient = ((MainActivity) getContext()).getService().getClient();
             var lobbyId = ((MainActivity) getContext()).getLobbyId();
@@ -55,6 +64,7 @@ public class InGameFragment extends Fragment {
             websocketClient.send(message);
             NavHostFragment.findNavController(InGameFragment.this)
                     .navigate(R.id.action_InGameFragment_to_FirstFragment);
+
         });
 
 
@@ -79,5 +89,36 @@ public class InGameFragment extends Fragment {
 
             websocketClient.send(message);
         });
+
+        sensorManager = (SensorManager) getActivity().getSystemService(Context.SENSOR_SERVICE);
+        sensor = sensorManager.getDefaultSensor(Sensor.TYPE_LIGHT); // Type_Light ist der int Wert 5
+    }
+
+
+
+    @Override
+    public void onSensorChanged(SensorEvent event) {
+        float x = event.values[0];
+
+        if (x < 40){
+            Log.e("Code", "sensor_light");
+            GameManager.getInstance().moveWormholes();
+        }
+    }
+
+    @Override
+    public void onAccuracyChanged(Sensor sensor, int i) {
+
+    }
+    @Override
+    public void onResume() {
+        super.onResume();
+        sensorManager.registerListener(this, sensor, SensorManager.SENSOR_DELAY_NORMAL);
+    }
+
+    @Override
+    public void onPause() {
+        super.onPause();
+        sensorManager.unregisterListener(this);
     }
 }
